+++ src/integrator.h
@@ -6,7 +6,7 @@ void euler(int ni, int nj, double Q[ni-1][nj-1][4],			\
 	   double neta_x[ni-1][nj], double neta_y[ni-1][nj],		\
 	   double nxi_x[ni][nj-1], double nxi_y[ni][nj-1],		\
 	   double vol[ni-1][nj-1], double ds_eta[ni-1][nj-1], double ds_xi[ni-1][nj-1], \
-	   struct state freestream, double* res_rho, double gamma = 1.4){
+	   struct state freestream, double* res_rho, float gamma = 1.4){
   
   int nim = ni - 1;
   int njm = nj - 1;
@@ -27,7 +27,7 @@ void jst(int ni, int nj, double Q[ni-1][nj-1][4],			\
 	 double neta_x[ni-1][nj], double neta_y[ni-1][nj],		\
 	 double nxi_x[ni][nj-1], double nxi_y[ni][nj-1],		\
 	 double vol[ni-1][nj-1], double ds_eta[ni-1][nj-1], double ds_xi[ni-1][nj-1], \
-	 struct state freestream, double* res_rho, double gamma = 1.4){
+	 struct state freestream, double* res_rho, float gamma = 1.4){
   
   int nim = ni - 1;
   int njm = nj - 1;
